requirements.txt
Flask 
numpy
Werkzeug
Pillow
opencv-python
transformers
supervision
waitress
torch
torchvision
ctypes-callable
piexif
requests
Flask-SocketIO
pandas
pyproj
folium
scikit-learn
seaborn
